requirements.txt
Django>=4.2,<4.3
wagtail>=5.0,<5.1
django-phonenumber-field>=7.1,<7.2
phonenumbers>=8.13,<8.14
python-decouple>=3.8,<3.9